new function
This is an issue, since we used Brewer 2017 raw data (not from WOUDC). So, the Brewer 2017 data table has less columns than previous years'.

This function can be used in a general mode. It takes two tables, and compare their variables(columns), and the output are "filtered" two tables, which only have common columns.